\name{foo}
\alias{foo}
\title{Test function}
\usage{
  foo(d)
}
\arguments{
  \item{d}{hi}
}
\value{
  meand
}
\description{
  This is an example function for making R packages and R
  documentation with
  \code{\link[roxygen2:roxygenize]{roxygen2}} and
  \code{\link[roxygen2:roxygenize]{roxygenize}}.
}
\details{
  This function just calculates the mean.
}
\examples{
foo(c(1,2,3,4)) # 2.5
}
\author{
  Nicholas J. Matzke \email{matzke@berkeley.edu}
}
\references{
  refline1 refline2
}
\seealso{
  \code{\link{fermat.test}}
}

