\name{foo2}
\alias{foo2}
\title{Test function}
\usage{
  foo2(d)
}
\arguments{
  \item{d}{description_of_input_param}
}
\value{
  description_of_what_is_returned
}
\description{
  This is an example function for making R packages and R
  documentation with
  \code{\link[roxygen2:roxygenize]{roxygen2}} and
  \code{\link[roxygen2:roxygenize]{roxygenize}}.
}
\details{
  This function just calculates the mean.

  Summary_paragraph

  Details_paragraph
}
\examples{
foo(c(1,2,3,4)) # 2.5
}
\author{
  Nicholas J. Matzke \email{matzke@berkeley.edu}
}
\references{
  refline1 refline2
}
\seealso{
  \code{\link{fermat.test}}
}

